fix(android): thread ids are timestamps, and the voice service must not claim the mic while idle
Chat thread ids are Date.now() values, so decoding them as Int clamped
every thread to 2147483647: the threads sheet crashed on duplicate keys
and the active thread always read as 0. Ids are Long end to end and the
top-level number results decode as Double first.

The foreground service asked for the microphone type whenever Bee was
thinking or speaking; Android 14+ throws unless RECORD_AUDIO is granted
and the app is recording. It now claims microphone only while listening
and a refused start no longer takes the app down.

+++ apps/android/app/src/main/kotlin/com/beegreat/app/voice/VoiceNoteController.kt
@@ -65,7 +65,7 @@ class VoiceNoteController(
   private var speechJob: Job? = null
   private var speechGeneration = 0
   private var spokenIds = HashSet<String>()
-  private var seededThread: Int? = null
+  private var seededThread: Long? = null
 
   fun start() {
     combine(agent.state, preferences.speakReplies) { agentState, speak -> agentState to speak }
@@ -173,7 +173,7 @@ class VoiceNoteController(
     }
   }
 
-  private fun considerSpeaking(thread: Int, status: AgentStatus, messages: List<FlueMessage>, speak: Boolean) {
+  private fun considerSpeaking(thread: Long, status: AgentStatus, messages: List<FlueMessage>, speak: Boolean) {
     if (seededThread != thread) {
       // Everything already in the transcript was said before; only new replies get read.
       seededThread = thread
